# Conflicts:
#	krrood/src/krrood/class_diagrams/class_diagram.py
#	krrood/src/krrood/class_diagrams/wrapped_field.py
#	krrood/src/krrood/ormatic/dao.py
#	krrood/src/krrood/ormatic/wrapped_table.py
#	pycram/src/pycram/orm/ormatic_interface.py
#	semantic_digital_twin/scripts/generate_orm.py
#	test/krrood_test/dataset/example_classes.py
#	test/krrood_test/dataset/ormatic_interface.py
#	test/krrood_test/test_ormatic/test_interface.py
#	test/semantic_digital_twin_test/test_orm/test_orm.py

+++ pycram/src/pycram/datastructures/grasp.py
@@ -15,7 +15,6 @@ from semantic_digital_twin.world_description.world_entity import Body
 from .dataclasses import Rotations
 from .enums import Grasp, AxisIdentifier, ApproachDirection, VerticalAlignment
 from .pose import PoseStamped, PyCramVector3
-from ..has_parameters import HasParameters
 from ..tf_transformations import quaternion_multiply
 from ..utils import translate_pose_along_local_axis
 
@@ -161,7 +160,7 @@ class GraspDescription:
 
         t = grasp_global.inverse().to_np() @ front_global.to_np()
 
-        return t[:3, 3].astype(int).tolist()
+        return t[:3, 3].astype(float).tolist()
 
     def grasp_orientation(self):
         """
